backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.3.4
google-generativeai==0.8.3
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11
supabase==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
pillow==10.3.0